requirements.txt
fastapi
uvicorn
torch
transformers
facenet-pytorch
pillow
numpy
python-multipart
